ci: use rangutopia for publishing
Replace the homegrown publish scripts with the rangutopia CLI. Versioning,
checking and publishing run as separate workflow steps, and a new
release-root action bumps the root and client packages, only when a public
library is being released. The bump commit stays unpushed so a failed
publish leaves nothing on the remote.

Also grant contents:write for the github releases, ignore release commits
in commitlint, declare missing workspace deps and bump all actions to v6.
