+++ .github/actions/upload-build-artifacts/action.yml
@@ -15,7 +15,7 @@ runs:
         echo "hashed=${hashed_name}">> $GITHUB_OUTPUT
       shell: 'bash'
     - name: Upload Build Artifacts
-      uses: actions/upload-artifact@v4
+      uses: actions/upload-artifact@v6
       with:
         name: ${{steps.hash_current_branch_name.outputs.hashed}}
         path: |
