+++ .github/workflows/checks.yaml
@@ -11,7 +11,7 @@ jobs:
   check-commits:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
         with:
           # Full history is required so check-conventional-commits can run
           # `git log origin/<base>..HEAD`.
@@ -26,14 +26,14 @@ jobs:
   build-libs:
     runs-on: ubuntu-latest
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Prepare
         uses: ./.github/actions/prepare
 
       # Cache Nx outputs (.nx/cache) and metadata (.nx/workspace-data) to optimize incremental builds
       - name: Cache nx computation cache
-        uses: actions/cache@v4
+        uses: actions/cache@v6
         with:
           path: |
             .nx/cache
@@ -59,9 +59,9 @@ jobs:
     runs-on: ubuntu-latest
     needs: build-libs
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
       - name: Download Build Information For Current Branch
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v6
         with:
           name: ${{needs.build-libs.outputs.hashed_branch}}
       - name: Prepare
@@ -78,20 +78,20 @@ jobs:
       REACT_APP_RANGO_API_KEY: ${{ vars.RANGO_API_KEY }}
 
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
 
       - name: Prepare
         uses: ./.github/actions/prepare
 
       - name: Download Build Information For Current Branch
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v6
         continue-on-error: true
         with:
           name: ${{needs.build-libs.outputs.hashed_branch}}
 
       # Read-only restore of the Nx build state to skip rebuilding shared dependencies
       - name: Restore nx computation cache
-        uses: actions/cache/restore@v4
+        uses: actions/cache/restore@v6
         with:
           path: |
             .nx/cache
@@ -148,13 +148,13 @@ jobs:
     runs-on: ubuntu-latest
     needs: build-libs
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
         with:
           ref: ${{ github.ref }}
       - name: Prepare
         uses: ./.github/actions/prepare
       - name: Download Build Information For Current Branch
-        uses: actions/download-artifact@v4
+        uses: actions/download-artifact@v6
         with:
           name: ${{needs.build-libs.outputs.hashed_branch}}
       - name: Build examples
