+++ commitlint.config.js
@@ -1,5 +1,8 @@
 const Configuration = {
   extends: ['@commitlint/config-conventional'],
+
+  // Release commits list every published package, far past `body-max-line-length`.
+  ignores: [(msg) => /^chore\(release\):/.test(msg)],
 };
 
 module.exports = Configuration;
