+++ package.json
@@ -41,7 +41,9 @@
     "dev:playground": "concurrently \"yarn dev:playground:dev-server\" \"yarn dev:playground:type-checker\"",
     "lint": "nx run-many --target=lint",
     "format": "nx run-many --target=format",
-    "publish": "node ./scripts/publish/command.mjs",
+    "publish:version": "rangutopia library version --deferred",
+    "publish:version:check": "rangutopia library version check",
+    "publish": "rangutopia library publish",
     "deploy": "node ./scripts/deploy/command.mjs",
     "sync-branch": "node ./scripts/sync-branch/command.mjs",
     "upgrade-all": "node ./scripts/upgrade-all/command.mjs --project ",
@@ -76,7 +78,6 @@
     "conventional-changelog-angular": "^8.0.0",
     "conventional-commits-filter": "^5.0.0",
     "conventional-commits-parser": "^6.2.0",
-    "conventional-recommended-bump": "^11.2.0",
     "esbuild": "^0.20.1",
     "esbuild-plugins-node-modules-polyfill": "^1.6.6",
     "eslint": "^9.19.0",
@@ -97,7 +98,7 @@
     "patch-package": "^8.0.0",
     "postinstall-postinstall": "^2.1.0",
     "prettier": "^2.7.1",
-    "rangutopia": "^0.3.1",
+    "rangutopia": "^1.0.0",
     "rimraf": "^5.0.1",
     "semver-parser": "^4.1.4",
     "typescript-eslint": "^8.21.0",
