--- scripts/common/constants.mjs
@@ -1,5 +1,0 @@
-/** A fixed subject for using in publish commit. */
-export const PUBLISH_COMMIT_SUBJECT = 'chore(release): publish';
-export const NPM_ORG_NAME = '@rango-dev';
-export const ROOT_VERSIONS_COMMIT_SUBJECT =
-  'chore(release): bump the repo and client versions';
