+++ scripts/common/errors.mjs
@@ -1,61 +1,5 @@
-export class GithubGetReleaseError extends Error {
-  name = 'GithubGetReleaseError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
-export class GithubReleaseNotFoundError extends Error {
-  name = 'GithubReleaseNotFoundError';
-  constructor(tag) {
-    super(`Couldn't find any github release for ${tag}`);
-  }
-}
-
-export class GithubCreateReleaseFailedError extends Error {
-  name = 'GithubCreateReleaseFailedError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
 export class GithubCommandError extends Error {
   name = 'GithubCommandError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
-export class NpmPackageNotFoundError extends Error {
-  name = 'NpmPackageNotFoundError';
-  constructor(packageName) {
-    super(`Couldn't find ${packageName} on NPM.`);
-  }
-}
-
-export class NpmGetPackageError extends Error {
-  name = 'NpmGetPackageError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
-export class NpmPublishError extends Error {
-  name = 'NpmPublishError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
-export class IncreaseVersionFailedError extends Error {
-  name = 'IncreaseVersionFailedError';
-  constructor(msg) {
-    super(msg);
-  }
-}
-
-export class UnableToProceedPublishError extends Error {
-  name = 'UnableToProceedPublishError';
   constructor(msg) {
     super(msg);
   }
