+++ scripts/common/npm.mjs
@@ -11,92 +11,6 @@ import {
 import { detectChannel } from './github.mjs';
 import { rootPath } from './path.mjs';
 
-/**
- * Publish a package using `yarn publish`
- *
- * @param {import('../common/utils.mjs').Package} pkg
- */
-export async function publishOnNpm(pkg) {
-  const channel = detectChannel();
-  const distTag = channel === 'prod' ? 'latest' : channel;
-  const output = await execa('yarn', [
-    'publish',
-    pkg.location,
-    '--tag',
-    distTag,
-  ])
-    .then(({ stdout }) => stdout)
-    .catch((error) => {
-      throw new NpmPublishError(error.stderr);
-    });
-
-  return output;
-}
-
-/**
- * Accept a package and send a request to npm registry for getting information.
- *
- * @param {import('./typedefs.mjs').Package} pkg
- * @returns {Promise<import('./typedefs.mjs').NpmVersions>}
- */
-export async function getNpmPackage(pkg) {
-  const packageName = pkg.name;
-  const headers = new Headers();
-  // This is to use less bandwidth unless we really need to get the full response.
-  // See https://github.com/npm/npm-registry-client#request
-  headers.append(
-    'Accept',
-    'application/vnd.npm.install-v1+json; q=1.0, application/json; q=0.8, */*'
-  );
-  const response = await fetch(
-    `https://registry.npmjs.org/${escapeName(packageName)}`,
-    {
-      headers,
-    }
-  ).catch((err) => {
-    const msg =
-      err.message || 'An error has occured when trying to get npm package.';
-    throw new NpmGetPackageError(msg);
-  });
-
-  const body = await response.json();
-
-  // A new package which never has been published on npm.
-  if (response.status === 404) {
-    throw new NpmPackageNotFoundError(packageName);
-  } else if (response.status > 300) {
-    const msg = `Package: ${packageName}, Status: ${response.status}, Body: ${body}`;
-    throw new NpmGetPackageError(msg);
-  }
-
-  const versions = {
-    next: body['dist-tags'].next || null,
-    prod: body['dist-tags'].latest || null,
-  };
-
-  return versions;
-}
-
-/**
- * Getting version from NPM
- * returns null if found anything.
- *
- * @param {import('./typedefs.mjs').Package} pkg
- * @returns {Promise<import('./typedefs.mjs').NpmVersions | null>}
- */
-export async function npmVersionFor(pkg) {
-  try {
-    const npmVersions = await getNpmPackage(pkg);
-    return npmVersions;
-  } catch (err) {
-    if (err instanceof NpmPackageNotFoundError) {
-      return null;
-    }
-
-    throw err;
-  }
-}
-
 export async function packagePath(project) {
   const { stdout: info } = await $`yarn workspaces info`.catch((error) => {
     throw new YarnError(`'yarn workspaces info' failed. \n ${error.stderr}`);
