+++ scripts/common/typedefs.mjs
@@ -8,48 +8,3 @@
  * @property {boolean} private - Whether the package is private or not.
  */
 
-/**
- * @typedef {Object} Release
- * @property {string} tagName - The name of the package followed by version.
- */
-
-/**
- * @typedef {Object} Tag
- * @property {string} tagName - The name of the package followed by version.
- */
-
-/**
- * @typedef {Object} PackageAndRelease
- * @property {Package} package
- * @property {Release | null} release
- */
-
-/**
- * @typedef {Object} PackageAndTag
- * @property {Package} package
- * @property {Tag | null} tag
- */
-
-/**
- * @typedef {Object} NpmVersions
- * @property {string} next - last next release version
- * @property {string} prod - last prod release version
- */
-
-/**
- * @typedef {Object} PackageAndNpmVersions
- * @property {Package} package
- * @property {NpmVersions | null} npm
- */
-
-/**
- * @typedef {Object} IncreaseVersionResult
- * @property {string} current - current version
- * @property {string} next - increased version
- */
-
-/**
- * @typedef {Object} CommitOptions
- * @property {boolean} shouldSkipCI - If true, appends `[skip ci]` to the commit message.
- * @property {boolean} shouldVerify - If true, appends `[skip ci]` to the commit message.
- */
