--- scripts/publish/typedefs.mjs
@@ -1,9 +1,0 @@
-/**
- * @typedef {Object} PackageState
- * @property {string | null} version
- * @property {string | null} githubRelease
- * @property {string | null} gitTag
- * @property {string | null} npmVersion
- * @example
- * { '@myorg/x-pkg': {version: "1.0.0", gitTag: "1.0.0", githubRelease: "1.0.0", npmVersion: "1.0.0"} }
- */
